+++ gitlabber/progress.py
@@ -13,7 +13,7 @@ class ProgressBar:
         if self.progress is None:
             self.start = time.time()
             self.progress = tqdm(total=total, unit="projects",
-                                 bar_format="{desc}: {percentage:.1f}%|{bar:100}| {n_fmt}/{total_fmt}{postfix}", desc=self.description, leave=False, disable=self.disabled)
+                                 bar_format="{desc}: {percentage:.1f}%|{bar:80}| {n_fmt}/{total_fmt}{postfix}", desc=self.description, leave=False, disable=self.disabled)
 
     def update_progress_length(self, added):
         if self.progress is not None:
@@ -29,7 +29,7 @@ class ProgressBar:
     def finish_progress(self):
         if self.progress is not None:
             self.progress.close()
-            end = time.time() - self.start
+            end = time.time()
             hours, rem = divmod(end-self.start, 3600)
             minutes, seconds = divmod(rem, 60)
             return "{:0>2}:{:0>2}:{:05.2f}".format(int(hours), int(minutes), seconds)