+++ CHANGELOG.md
@@ -1,3 +1,109 @@
-1.1.0 - Add support for HTTP clone via CLI argument or ENV variable
+# Changelog
 
-1.0.9 - First published version on PyPi
+<!--next-version-placeholder-->
+## [1.1.8]  - 24/03/2021
+
+### Added
+### Changed
+ - Changelog standartization
+### Deprecated
+### Removed
+### Fixed
+ - Fix recursive flag
+### Security
+
+
+## [1.1.7]  - 15/03/2021
+
+### Added
+### Changed
+ - Move to GH actions
+### Deprecated
+### Removed
+### Fixed
+### Security
+## [1.1.6] - 11/03/2021
+
+### Added
+### Changed
+- Update depedencies
+### Deprecated
+### Removed
+### Fixed
+### Security
+## [1.1.4] - 10/03/2021
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- Fixes to naming strategy
+### Security
+## [1.1.3] - 08/03/2021
+### Added
+- Support git submodules
+
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+## [1.1.2] - 22/02/2021
+### Added
+- Support different local folder naming strategy (project path/name)
+- support CA bundles 
+
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+## [1.1.1] - 01/05/2021
+### Added
+ - Added concurrency 
+
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- Fix for pattern matching filtering out parents with relevant children
+- Fix Gitlab groups API change not returning subgroups
+### Security
+## [1.1.0] - - 11/05/2020
+### Added
+- Add support for HTTP clone via CLI argument or ENV variable
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+ ## [1.0.9] - 10/04/2020
+### Added
+
+- Documentation Changes
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+ ## [1.0.8] - 28/03/2020
+### Added
+
+- First published version on PyPi
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+[unreleased]: https://github.com/ezbz/gitlabber/compare/v1.1.8...HEAD
+[1.1.8]: https://github.com/ezbz/gitlabber/compare/v1.1.7...v1.1.8
+[1.1.7]: https://github.com/ezbz/gitlabber/compare/v1.1.6...v1.1.7
+[1.1.6]: https://github.com/ezbz/gitlabber/compare/v1.1.4...v1.1.6
+[1.1.4]: https://github.com/ezbz/gitlabber/compare/v1.1.3...v1.1.4
+[1.1.3]: https://github.com/ezbz/gitlabber/compare/v1.1.2...v1.1.3
+[1.1.2]: https://github.com/ezbz/gitlabber/compare/v1.1.1...v1.1.2
+[1.1.1]: https://github.com/ezbz/gitlabber/compare/v1.1.0...v1.1.1
+[1.1.0]: https://github.com/ezbz/gitlabber/compare/v1.0.9...v1.1.0
+[1.0.9]: https://github.com/ezbz/gitlabber/compare/v1.0.8...v1.0.9
+[1.0.8]: https://github.com/ezbz/gitlabber/releases/tag/v1.0.8